Initial commit of code ported from Django (to Google App Engine).
This is the first commit of the GAE code, ported from Django. Included
only the Python files from the boto module. GAE only runs pure Python
code, so the binaries were omitted. Note that this causes some of the
boto library to be nonfunctional, but those parts are not used in this
code.