(Check out next revision, not this one!) Coded mechanism to keep AWS access key id and secret access key private from other developers on this project.
To configure the project in a cloned repository for debugging and
testing, enter your Amazon Web Services (AWS) access key id and secret
access key into the template provided in the file settings.py. Update
to the next revision to receive the latest .ignore file that ignores
the settings.py files, to protect your AWS credentials.